Refactor code that gets posts
Getting posts relied on having a userId param
within the URI path which made it difficult to
access all posts regardless of author. The
getPosts function has now been changed to get all
the posts first and then filter the list down
depending on which params or queries are provided.
This simplifies the database, reduces the amount
of extra functions in the postsController and also
enables accessing of all posts from all authors.
Getting posts ALL and posts SINGULAR is also
decoupled from the user route, but will also still
work with the user route too.