Make middleware check token for author status
Previously the user object stored on req was being
checked to verify if the user was an author or not
This wasn't what I was aiming for. Instead now the
users author status is stored on the JWT token
and the JWT token's payload is checked by the
verifyAuthor middleware as opposed to checking the
req.user object.