Protect comments PUT and DELETE routes
In order to update or delete comments users must
be signed in. To delete a specific resource a
userid must be provided as a parameter as well as
an id for the comment. The userid provided as a
parameter must also match the id stored in the
JWT. This ensures the user can only delete or
update posts that belong to themselves.