Change add post to include user name as author
When posts are retrieved on the client side it
exposes the user's ID but not the author's name of
the post. In order to retrieve the post's author's
name it will become complex to make multiple calls
to the database. Instead I've opted to add a
author field to the Post schema. I'll move towards
refactoring code to change exposing the user's
ID on the front end where it's not required.